Correct DPI by all TCastleWindow backends and TCastleControl
This includes TCastleWindow backends GTK, WinAPI, X11, LCL (Android
and iOS were already OK).

+++ src/3d/castlecameras.pas
@@ -2557,7 +2557,7 @@ var
   ModsDown: TModifierKeys;
   DoZooming, DoMoving: boolean;
   MoveDivConst: Single;
-  Dpi: Integer;
+  Dpi: Single;
 
   function DragRotation: TQuaternion;
